api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19
httpx==0.28.1
minio==7.2.12
pymilvus==2.4.9
pydantic==2.10.3
pydantic-settings==2.7.0
apscheduler==3.10.4
python-docx==1.1.2
pdfplumber==0.11.4
pypdf==5.1.0
langchain==0.3.13
langchain-community==0.3.13
langgraph==0.2.60
FlagEmbedding==1.2.11
sentence-transformers==3.3.1
